moved Scheduler to extension
